% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Proxy.r
\name{is_proxy}
\alias{is_proxy}
\title{Lookup for IP address proxy information}
\usage{
is_proxy(ip)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}
}
\value{
Return a digit value: -1 (errors), 0 (not a proxy), 1 (a proxy), and 2 (a data center IP address).
}
\description{
Check whether if an IP address was a proxy.
}
\examples{
\donttest{
is_proxy("8.8.8.8")
}

}
