% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Proxy.r
\name{lookup_web_service}
\alias{lookup_web_service}
\title{Lookup for IP address proxy information using IP2Proxy web service.}
\usage{
lookup_web_service(api_key, ip, package = "PX1")
}
\arguments{
\item{api_key}{IP2Proxy web service API key}

\item{ip}{IPv4 or IPv6 address}

\item{package}{Package to use for IP2Proxy web service.}
}
\value{
Return all the proxy information about the IP address
}
\description{
Find the country, region, city, ISP, domain name, usage types, asn, as name, last seen, threat type and provider. The return values will be depending on the IP2Proxy web service package used.
}
\examples{
\dontrun{
lookup_web_service("1.0.241.135","PX1")
}

}
