% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Proxy.r
\name{open}
\alias{open}
\title{Load IP2Proxy BIN data}
\usage{
open(bin_location)
}
\arguments{
\item{bin_location}{Absolute path of IP2Proxy BIN data}
}
\description{
Load the IP2Proxy BIN data for lookup. Free IP2Proxy LITE data available for download at <https://lite.ip2location.com/>
}
\examples{
\donttest{
open("~/IP-COUNTRY.BIN")
}

}
