% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Proxy_plotmap.R
\name{plot_map}
\alias{plot_map}
\title{Plot map using IP2Location data.}
\usage{
plot_map(ips)
}
\arguments{
\item{ips}{A vector of IP addresses to be plot on}
}
\description{
Plot the country on the map based on IP addresses and its IP2Location country data.
}
\examples{
\dontrun{
plot_map(c("1.0.241.135", "1.2.3.4"))
}

}
