% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Proxy.r
\name{get_all}
\alias{get_all}
\title{Lookup for IP address proxy information}
\usage{
get_all(ip)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}
}
\value{
Return all the proxy information about the IP address
}
\description{
Find the country, region, city, ISP, domain name, usage types, asn, as name, last seen, threat type and provider. The return values will be depending on the BIN data loaded.
}
\examples{
\dontrun{
get_all("1.0.241.135")
}

}
